Modularize args.rs into submodules for maintainability
Split src/cli/args.rs (533 lines) into domain-based submodules:
- args/mod.rs: Main CLI struct, Commands enum, smaller command enums
- args/issues.rs: IssueCommands enum
- args/views.rs: ViewCommands enum
- args/goals.rs: GoalCommands enum

Follows screaming architecture pattern from recent modularization commits.
